requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.3
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
